run workflow on pull_request
iiuc, the tests are currently not running for PRs and this will run the builds for PRs. considering how many builds there are in this workflow maybe you're worried about how much build time that can be. we could a workflow just for PRs that runs on only one llvm version. idk